+++ nqueens-server/src/main/java/Server.java
@@ -1,0 +1,130 @@
+package nqueens;
+
+import com.sun.net.httpserver.Headers;
+import com.sun.net.httpserver.HttpServer;
+import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.net.InetSocketAddress;
+import java.net.URI;
+import java.net.URLDecoder;
+import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
+// curl localhost:8081/func1
+
+public class Server {
+    private static final String HOSTNAME = "localhost";
+    private static final int PORT = 8081;
+    private static final int BACKLOG = 1;
+
+    private static final String HEADER_ALLOW = "Allow";
+    private static final String HEADER_CONTENT_TYPE = "Content-Type";
+
+    private static final Charset CHARSET = StandardCharsets.UTF_8;
+
+    private static final int STATUS_OK = 200;
+    private static final int STATUS_METHOD_NOT_ALLOWED = 405;
+
+    private static final int NO_RESPONSE_LENGTH = -1;
+
+    private static final String METHOD_GET = "GET";
+    private static final String METHOD_OPTIONS = "OPTIONS";
+    private static final String ALLOWED_METHODS = METHOD_GET + "," + METHOD_OPTIONS;
+
+    public static void main(final String... args) throws IOException {
+        final HttpServer server = HttpServer.create(new InetSocketAddress(HOSTNAME, PORT), BACKLOG);
+        server.createContext("/nqueens", he -> {
+            try {
+                final Headers headers = he.getResponseHeaders();
+                final String requestMethod = he.getRequestMethod().toUpperCase();
+                switch (requestMethod) {
+                    case METHOD_GET:
+                        final Map<String, List<String>> requestParameters = getRequestParameters(he.getRequestURI());
+			// boolean noThreeInLineConstraint = requestParameters.containsKey("l");
+			String responseBody = "";
+
+			if (requestParameters.containsKey("n")) {
+			    List<String> rs = requestParameters.get("n");
+			    if (rs.size() == 1) {
+				try {
+				    int n = Integer.parseInt(rs.get(0));
+				    // BruteForceSolver ns = new BruteForceSolver(n,noThreeInLineConstraint);
+				    BruteForceSolver ns = new BruteForceSolver(n, false);
+				    responseBody = ns.solutionsToJSON();
+				} catch (NumberFormatException e) {
+				    responseBody = "{\"error\":\"n should be an integer\"}";
+				}
+			    } else {
+				responseBody = "{\"error\":\"n should be an integer\"}";
+			    }
+			} else {
+			    responseBody = "{\"error\":\"this query requires a parameter n=<int>\"}";
+			}
+
+                        headers.set(HEADER_CONTENT_TYPE, String.format("application/json; charset=%s", CHARSET));
+                        final byte[] rawResponseBody = responseBody.getBytes(CHARSET);
+                        he.sendResponseHeaders(STATUS_OK, rawResponseBody.length);
+                        he.getResponseBody().write(rawResponseBody);
+                        break;
+
+		    case METHOD_OPTIONS:
+                        headers.set(HEADER_ALLOW, ALLOWED_METHODS);
+                        he.sendResponseHeaders(STATUS_OK, NO_RESPONSE_LENGTH);
+                        break;
+
+		    default:
+                        headers.set(HEADER_ALLOW, ALLOWED_METHODS);
+                        he.sendResponseHeaders(STATUS_METHOD_NOT_ALLOWED, NO_RESPONSE_LENGTH);
+                        break;
+                }
+            } finally {
+                he.close();
+            }
+        });
+        server.start();
+    }
+
+    private static Map<String, List<String>> getRequestParameters(final URI requestUri) {
+        final Map<String, List<String>> requestParameters = new LinkedHashMap<>();
+        final String requestQuery = requestUri.getRawQuery();
+        if (requestQuery != null) {
+            final String[] rawRequestParameters = requestQuery.split("[&;]", -1);
+            for (final String rawRequestParameter : rawRequestParameters) {
+                final String[] requestParameter = rawRequestParameter.split("=", 2);
+                final String requestParameterName = decodeUrlComponent(requestParameter[0]);
+                requestParameters.putIfAbsent(requestParameterName, new ArrayList<>());
+                final String requestParameterValue = requestParameter.length > 1 ? decodeUrlComponent(requestParameter[1]) : null;
+                requestParameters.get(requestParameterName).add(requestParameterValue);
+            }
+        }
+        return requestParameters;
+    }
+
+    private static String decodeUrlComponent(final String urlComponent) {
+        try {
+            return URLDecoder.decode(urlComponent, CHARSET.name());
+        } catch (final UnsupportedEncodingException ex) {
+            throw new InternalError(ex);
+        }
+    }
+}
+
+// import org.json.JSONException;
+// import org.json.JSONObject;
+// import java.io.BufferedReader;
+// import java.io.DataOutputStream;
+// import java.io.FileNotFoundException;
+// import java.io.InputStreamReader;
+// import java.net.HttpURLConnection;
+// import java.net.URL;
+// import java.util.Map;
+
+// public class Server {
+//     public static void main(String[] args) {
+
+//     }
+// }